+++ pkg/bootstrap/bootstrap.go
@@ -22,7 +22,7 @@ import (
 	"strmprivacy/strm/pkg/entity/kafka_user"
 	"strmprivacy/strm/pkg/entity/key_stream"
 	"strmprivacy/strm/pkg/entity/keylinks"
-	"strmprivacy/strm/pkg/entity/member"
+	"strmprivacy/strm/pkg/entity/user"
 	"strmprivacy/strm/pkg/entity/organization"
 	"strmprivacy/strm/pkg/entity/policy"
 	"strmprivacy/strm/pkg/entity/project"
@@ -73,7 +73,7 @@ func SetupServiceClients(accessToken *string) {
 	installation.SetupClient(clientConnection, ctx)
 	account.SetupClient(clientConnection, ctx)
 	project.SetupClient(clientConnection, ctx)
-	member.SetupClient(clientConnection, ctx)
+	user.SetupClient(clientConnection, ctx)
 	organization.SetupClient(clientConnection, ctx)
 	data_subject.SetupClient(clientConnection, ctx)
 	keylinks.SetupClient(clientConnection, ctx)