+++ pkg/cmd/manage.go
@@ -4,7 +4,7 @@ import (
 	"github.com/spf13/cobra"
 	"strmprivacy/strm/pkg/auth"
 	"strmprivacy/strm/pkg/common"
-	"strmprivacy/strm/pkg/entity/member"
+	"strmprivacy/strm/pkg/entity/user"
 	"strmprivacy/strm/pkg/entity/project"
 )
 
@@ -17,5 +17,5 @@ var ManageCmd = &cobra.Command{
 
 func init() {
 	ManageCmd.AddCommand(project.ManageCmd())
-	ManageCmd.AddCommand(member.ManageCmd())
+	ManageCmd.AddCommand(user.ManageCmd())
 }