+++ pkg/entity/batch_job/cmd.go
@@ -1,33 +1,21 @@
 package batch_job
 
 import (
-	"fmt"
 	"github.com/spf13/cobra"
-	"strmprivacy/strm/pkg/common"
 	"strmprivacy/strm/pkg/entity/policy"
 )
 
 const (
-	batchJobsFileFlagName = "file"
+	batchJobFileFlagName = "file"
+	batchJobTypeFlagName = "type"
+	encryptionType = "encryption"
+	microAggregationType = "micro-aggregation"
 )
-
-var longDoc = `
-A Batch Job reads all events from a Data Connector and writes them to one or more Data Connectors,
-applying our privacy algorithm as defined by the job's configuration file.
-
-A [Data Connector](docs/04-reference/01-cli-reference/` + fmt.Sprint(common.RootCommandName) + `/create/data-connector.md) is a configuration
-entity that comprises a location (GCS bucket, AWS S3 bucket, ...) and associated credentials.
-
-A Data Connector must be created *before* you can create a batch job that uses it.
-
-### Usage
-`
 
 func DeleteCmd() *cobra.Command {
 	batchJob := &cobra.Command{
-		Use:               "batch-job [id ...]",
-		Short:             "Delete on or more Batch Jobs by id",
-		Long:              longDoc,
+		Use:               "batch-job (id ...)",
+		Short:             "Delete one or more Batch Jobs by id",
 		DisableAutoGenTag: true,
 		PreRun: func(cmd *cobra.Command, args []string) {
 			printer = configurePrinter(cmd)
@@ -47,9 +35,8 @@ func DeleteCmd() *cobra.Command {
 
 func GetCmd() *cobra.Command {
 	return &cobra.Command{
-		Use:               "batch-job [id]",
+		Use:               "batch-job (id)",
 		Short:             "Get a Batch Job by id",
-		Long:              longDoc,
 		DisableAutoGenTag: true,
 		PreRun: func(cmd *cobra.Command, args []string) {
 			printer = configurePrinter(cmd)
@@ -65,7 +52,6 @@ func ListCmd() *cobra.Command {
 	return &cobra.Command{
 		Use:               "batch-jobs",
 		Short:             "List Batch Jobs",
-		Long:              longDoc,
 		DisableAutoGenTag: true,
 		PreRun: func(cmd *cobra.Command, args []string) {
 			printer = configurePrinter(cmd)
@@ -81,6 +67,7 @@ func CreateCmd() *cobra.Command {
 		Use:               "batch-job",
 		Short:             "Create a Batch Job",
 		Long:              longDoc,
+		Example:           example,
 		DisableAutoGenTag: true,
 		PreRun: func(cmd *cobra.Command, args []string) {
 			printer = configurePrinter(cmd)
@@ -93,11 +80,12 @@ func CreateCmd() *cobra.Command {
 
 	flags := batchJob.Flags()
 
-	flags.StringP(batchJobsFileFlagName, "F", "",
-		`The path to the JSON file containing the batch job configuration`)
+	flags.StringP(batchJobFileFlagName, "F", "",
+		`the path to the JSON file containing the batch job configuration`)
+	flags.StringP(batchJobTypeFlagName, "T", "encryption",
+		`the type of batch job (encryption, micro-aggregation), defaults to encryption`)
 	policy.SetupFlags(batchJob, flags)
-	err := batchJob.MarkFlagRequired(batchJobsFileFlagName)
-	common.CliExit(err)
+	_ = batchJob.MarkFlagRequired(batchJobFileFlagName)
 
 	return batchJob
 }