+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [3.3.0](https://github.com/strmprivacy/cli/compare/v3.2.0...v3.3.0) (2022-12-07)
+
+
+### Features
+
+* **strm-1882:** add micro-aggregation batch job & FieldMetaData support ([#128](https://github.com/strmprivacy/cli/issues/128)) ([c760872](https://github.com/strmprivacy/cli/commit/c7608726b68497a755013bdbe25f61a174de728a))
+
+# [3.2.0](https://github.com/strmprivacy/cli/compare/v3.1.1...v3.2.0) (2022-12-06)
+
+
+### Features
+
+* **strm-1393:** remove sts references and STRMline docs ([#130](https://github.com/strmprivacy/cli/issues/130)) ([67ea503](https://github.com/strmprivacy/cli/commit/67ea50343a6c7ee12ddc491eda42f917493e6230))
+
 ## [3.1.1](https://github.com/strmprivacy/cli/compare/v3.1.0...v3.1.1) (2022-11-29)